Add additional context to --debug log messages
In the `main()` method, in debug mode, we print out the argument
dictionaries a few times but without context for which one is which.
This commit adds an extra line of debug logging each time that indicates
from which point in the process the args are originating. Hopefully this
will expedite any arg parsing debugging.

I've also removed one line of unused code that created an intermediate
object for the argparse object.